pyproject.toml
[project]
name = "resell-bot"
version = "0.1.0"
description = "Bot de détection de bonnes affaires sur les sites de vente/rachat de livres"
requires-python = ">=3.12"
dependencies = [
    "httpx[http2]>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "apscheduler>=3.10,<4",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
